# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedGama Telehealth API",
...
        "contact": {
            "name": "MedGama Dev Team",
            "email": "dev@medgama.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Local Development"
        },
        {
            "url": "https://api.medgama.com/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/chat/conversations": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "List conversations (denormalized last_message, unread_count)",
                "operationId": "bd436631aa9db291379544cbcadec892",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated conversations with other_user, last_message, unread_count"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Start or retrieve a 1:1 conversation",
                "description": "Business Rule: A patient can only start a conversation with a doctor if they have at least one confirmed/completed appointment together. Doctors, clinic owners, and admins bypass this restriction.",
                "operationId": "a87623873795097383e860ebb9a9a148",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_id"
                                ],
                                "properties": {
                                    "recipient_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Existing conversation returned"
                    },
                    "201": {
                        "description": "New conversation created"
                    },
                    "403": {
                        "description": "No confirmed/completed appointment with this doctor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/chat/conversations/{conversation}/messages": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "List messages (newest first, sender eager-loaded)",
                "operationId": "b8aa74c6ed67873e7eda5ebf0b5a107a",
                "parameters": [
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated messages (ChatMessageResource)"
                    },
                    "403": {
                        "description": "Not a participant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Send message (text or file). Broadcasts message.sent event via WebSocket.",
                "operationId": "d90e88080122e58bd6b69cb39c086d01",
                "parameters": [
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "Required if no attachment",
                                        "type": "string"
                                    },
                                    "attachment": {
                                        "description": "Image or document, max 20MB. Type auto-detected from MIME.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent (ChatMessageResource)"
                    },
                    "403": {
                        "description": "Not a participant"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/chat/conversations/{conversation}/read": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Bulk mark all unread messages as read (single UPDATE query)",
                "operationId": "19c50eb0b048a75bcf9519bb42c5dbfb",
                "parameters": [
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages marked as read with count"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/chat/unread-count": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Total unread message count across all conversations",
                "operationId": "eb5e1a9556b7447ff2f5bfe38b0a98e8",
                "responses": {
                    "200": {
                        "description": "Unread count"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "description": "╔══════════════════════════════════════════════════════════════════╗\n║  MedGama API — OpenAPI 3.0 Specification                       ║\n║  GDPR / KVKK / HIPAA Compliant Telehealth Platform             ║\n╚══════════════════════════════════════════════════════════════════╝",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "You are not authorized."
                    },
                    "code": {
                        "type": "string",
                        "example": "FORBIDDEN"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}